+++ packages/at_client/lib/src/client/local_secondary.dart
@@ -104,8 +104,10 @@ class LocalSecondary implements Secondary {
   LocalSecondary(
     this._atClient, {
     this.keyStore,
+    AtSyncQueue? syncQueue,
     void Function(DataEvent)? onEvent,
-  }) : _onEvent = onEvent {
+  })  : _syncQueue = syncQueue,
+        _onEvent = onEvent {
     _logger = AtSignLogger('LocalSecondary (${_atClient.getCurrentAtSign()})');
     keyStore ??= _atClient.persistenceBundle?.keyValueStore;
   }